client: refactor and reformat
