+++ client/src/components/types.ts
@@ -39,3 +39,21 @@ export function createConfig(rawData: any): Config {
         sleep_time: rawData.sleep_time,
     };
 }
+
+export interface DailyStat {
+    date: Date;
+    total_coffee: number;
+    avg_break_interval: number | null;
+    avg_water_interval: number | null;
+    avg_mood: string;
+}
+
+export function createDailyStat(rawData: any) {
+    return {
+        date: new Date(rawData.date),
+        total_coffee: rawData.total_coffee,
+        avg_break_interval: rawData.avg_break_interval,
+        avg_water_interval: rawData.avg_water_interval,
+        avg_mood: rawData.avg_mood,
+    };
+}
