+++ client/src/utils/helpers.ts
@@ -1,0 +1,14 @@
+export function formatDateTime(date: Date | null): string {
+    if (!date || !isFinite(date.getTime()) || date.getFullYear() <= 1) {
+        return "—";
+    }
+    return new Intl.DateTimeFormat(undefined, {
+        month: "short",
+        day: "numeric",
+        year: "numeric",
+        hour: "2-digit",
+        minute: "2-digit",
+        hour12: false,
+        timeZoneName: "short",
+    }).format(date);
+}
